Summarize DTUrtle results

Summarize the key results of the DTUrtle analysis to a gene-level data frame.

create_dtu_table(
  dturtle,
  add_gene_metadata = list(pct_gene_expr = "exp_in"),
  add_tx_metadata = list(max_pct_tx_expr = c("exp_in", max))
)

Arguments

dturtle
dturtle result object of posthoc_and_stager().

add_gene_metadata
A list of columns of the object's meta_table_gene, the gene-level meta data table. Names can be specified, which are used as the column names in the final output.

add_tx_metadata
A list of tuples for the object's meta_table_tx, the transcript-level meta data table. The tuples must consist of the name of the column in meta_table_tx and a gene-level summarization function. This function shall summarize the transcript-level information in such a way, that only one value for each gene is returned. Names can be specified, which are used as the column names in the final output.

Value

An extended dturtle object, including the added dtu_table. The dtu_table contains key statistics for all significant DTU genes. By default, the dtu_table contains the following columns:

"gene_ID": The used gene identifiers (gene name or id)

"gene_qvalue": Multiple testing corrected p-value (a.k.a. q-value) comparing all transcripts together between the two groups ("gene level").

"minimal_tx_value": The minimal multiple testing corrected p-value from comparing all transcripts individually between the two groups ("transcript level"). I.e. the q-value of the most significant transcript.

"number_tx": The number of analyzed transcripts for the specific gene.

"number_significant_tx": The number of significant transcripts from the 'transcript level' analysis.

"max(Group1-Group2")": Maximal proportional difference between the two comparisons groups. The difference is computed by subtracting the fitted mean of Group2 from the fitted mean of Group1 (Group1-Group2).

Additional columns from the meta_table_gene or gene-level summarized columns of meta_table_tx can optionally be carried over.

Details

This function provides an easy interface to summarize the key DTUrtle results together with user-defined meta data columns to a gene-level data frame.

See also