Return the maximal absolute difference for all transcripts of the provided gene

Return the maximal absolute difference for all transcripts of the provided gene

getmax(gID, dturtle)

Arguments

gID
gene-identifier

dturtle
dturtle object

Value

The maximal absolute difference value