Summarize DTUrtle results

Summarize the key results of the DTUrtle analysis to a gene-level data frame.

create_dtu_table(
  dturtle,
  add_gene_metadata = list(pct_gene_expr = "exp_in"),
  add_tx_metadata = list(max_pct_tx_expr = c("exp_in", max))
)

Arguments

dturtle
dturtle result object of posthoc_and_stager().

add_gene_metadata
A list of columns of the object's meta_table_gene, the gene-level meta data table. Names can be specified, which are used as the column names in the final output.

add_tx_metadata
A list of tuples for the object's meta_table_tx, the transcript-level meta data table. The tuples must consist of the name of the column in meta_table_tx and a gene-level summarization function. This function shall summarize the trancript-level information in such a way, that only one value for each gene is returned. Names can be specified, which are used as the column names in the final output.

Value

An extended dturtle object, including the added dtu_table.

Details

This function provides an easy interface to summarize the key DTUrtle results together with user-defined meta data columns to a gene-level data frame.

See also